test: move code
